patch version number in setup.py
The version number in the setup.py file does not match the version of the
source code.  This leads to a mismatch between the package version and the one
recorded in the .dist-info directory, Keras_Preprocessing-1.0.1.dist-info
which should be Keras_Preprocessing-1.0.2.dist-info.

The included patch correct the version number in the setup.py file.